+++ src/Ext/WarheadType/Detonate.cpp
@@ -33,7 +33,7 @@ void WarheadTypeExt::ExtData::Detonate(TechnoClass* pOwner, HouseClass* pHouse, 
 	{
 		if (this->BigGap)
 		{
-			for (auto pOtherHouse : *HouseClass::Array)
+			for (auto pOtherHouse : HouseClass::Array)
 			{
 				if (pOtherHouse->IsControlledByHuman() && // Not AI
 					!pOtherHouse->IsObserver() &&         // Not Observer
@@ -47,7 +47,7 @@ void WarheadTypeExt::ExtData::Detonate(TechnoClass* pOwner, HouseClass* pHouse, 
 		}
 
 		if (this->SpySat)
-			MapClass::Instance->Reveal(pHouse);
+			MapClass::Instance.Reveal(pHouse);
 
 		if (this->TransactMoney)
 		{
@@ -65,7 +65,7 @@ void WarheadTypeExt::ExtData::Detonate(TechnoClass* pOwner, HouseClass* pHouse, 
 			int index = GeneralUtils::ChooseOneWeighted(ScenarioClass::Instance->Random.RandomDouble(), &this->SpawnsCrate_Weights);
 
 			if (index < static_cast<int>(this->SpawnsCrate_Types.size()))
-				MapClass::Instance->PlacePowerupCrate(CellClass::Coord2Cell(coords), this->SpawnsCrate_Types.at(index));
+				MapClass::Instance.PlacePowerupCrate(CellClass::Coord2Cell(coords), this->SpawnsCrate_Types.at(index));
 		}
 
 		for (const int swIdx : this->LaunchSW)
@@ -209,7 +209,7 @@ void WarheadTypeExt::ExtData::ApplyBuildingUndeploy(TechnoClass* pTarget)
 
 	const auto pType = pBuilding->Type;
 
-	if (!pType->UndeploysInto || (pType->ConstructionYard && !GameModeOptionsClass::Instance->MCVRedeploy))
+	if (!pType->UndeploysInto || (pType->ConstructionYard && !GameModeOptionsClass::Instance.MCVRedeploy))
 		return;
 
 	auto cell = pBuilding->GetMapCoords();
@@ -270,7 +270,7 @@ void WarheadTypeExt::ExtData::ApplyBuildingUndeploy(TechnoClass* pTarget)
 		cell.Y += static_cast<short>(14 * sin(radian));
 
 		// Find a location where the conyard can be deployed
-		const auto newCell = MapClass::Instance->NearByLocation(cell, pType->UndeploysInto->SpeedType, -1, pType->UndeploysInto->MovementZone,
+		const auto newCell = MapClass::Instance.NearByLocation(cell, pType->UndeploysInto->SpeedType, -1, pType->UndeploysInto->MovementZone,
 			false, (width + 2), (height + 2), false, false, false, false, CellStruct::Empty, false, false);
 
 		// If it can find a more suitable location, go to the new one
@@ -278,7 +278,7 @@ void WarheadTypeExt::ExtData::ApplyBuildingUndeploy(TechnoClass* pTarget)
 			cell = newCell;
 	}
 
-	if (const auto pCell = MapClass::Instance->TryGetCellAt(cell))
+	if (const auto pCell = MapClass::Instance.TryGetCellAt(cell))
 		pBuilding->SetArchiveTarget(pCell);
 
 	pBuilding->Sell(1);
@@ -493,7 +493,7 @@ void WarheadTypeExt::ExtData::InterceptBullets(TechnoClass* pOwner, WeaponTypeCl
 	}
 	else
 	{
-		for (auto const pBullet : *BulletClass::Array)
+		for (auto const pBullet : BulletClass::Array)
 		{
 			if (pBullet->Location.DistanceFrom(coords) > cellSpread * Unsorted::LeptonsPerCell)
 				continue;